refactor(cl): convert AudioTestCase to mixin
